+++ Kochbuch/src/app/authentication/token-service.ts
@@ -7,7 +7,10 @@ import { Injectable } from '@angular/core';
  */
 @Injectable()
 export class TokenService {
+    constructor(){
+        console.log("tokenservice created");
+    }
     private token;
     public get Token() { return this.token }
-    public set Token(value: string) { this.token = value; }
+    public set Token(value: string) { this.token = value;console.log("token set"); }
 }
